Minor change: support for multiple indicators for the same objective

+++ src/module/siadap/domain/scoring/IScoring.java
@@ -1,6 +1,8 @@
 package module.siadap.domain.scoring;
 
+import java.math.BigDecimal;
+
 public interface IScoring {
 
-    public int getPoints();
+    public BigDecimal getPoints();
 }
